fix: Consolidate common exports between base package and compat package.
